app/requirements.txt
gunicorn
Flask
Flask-SQLAlchemy
flask-restx
SQLAlchemy
marshmallow
pyjwt
